// src/app/application/model/user.data.ts
export interface UserData {
  cedula: string;
  nombre: string;
  apellidos: string;
  sexo: string;
  lugar_de_nacimiento_1: string;
  lugar_de_nacimiento_2: string;
  fecha_de_nacimiento: string;
  nacionalidad: string;
  padre_cedula: string;
  padre_nombre_completo: string;
  madre_cedula: string;
  madre_nombre_completo: string;
  expedida: string;
  expira: string;
  barcode: string;
}